feat(user): Refactor user management and identity provider
Closes gravitee-io/issues#1595
