+++ src/test/java/io/gravitee/repository/config/mock/AbstractRepositoryMock.java
@@ -1,0 +1,46 @@
+/**
+ * Copyright (C) 2015 The Gravitee team (http://gravitee.io)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.gravitee.repository.config.mock;
+
+import org.mockito.Mockito;
+import org.springframework.beans.factory.FactoryBean;
+
+/**
+ * @author David BRASSELY (david.brassely at graviteesource.com)
+ * @author GraviteeSource Team
+ */
+public abstract class AbstractRepositoryMock<R> implements FactoryBean<R> {
+
+    private final Class<R> repoClass;
+
+    protected AbstractRepositoryMock(final Class<R> repoClass) {
+        this.repoClass = repoClass;
+    }
+
+    @Override
+    public R getObject() throws Exception {
+        R mock = Mockito.mock(repoClass);
+        prepare(mock);
+        return mock;
+    }
+
+    @Override
+    public Class<?> getObjectType() {
+        return repoClass;
+    }
+
+    abstract void prepare(R repository) throws Exception;
+}
