+++ src/test/java/io/gravitee/repository/config/mock/ApplicationRepositoryMock.java
@@ -1,0 +1,133 @@
+/**
+ * Copyright (C) 2015 The Gravitee team (http://gravitee.io)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.gravitee.repository.config.mock;
+
+import io.gravitee.repository.management.api.ApplicationRepository;
+import io.gravitee.repository.management.model.Application;
+import io.gravitee.repository.management.model.ApplicationStatus;
+import org.mockito.ArgumentMatcher;
+
+import java.util.HashSet;
+import java.util.Optional;
+import java.util.Set;
+
+import static io.gravitee.repository.utils.DateUtils.parse;
+import static java.util.Arrays.asList;
+import static java.util.Collections.singleton;
+import static java.util.Collections.singletonList;
+import static java.util.Optional.empty;
+import static java.util.Optional.of;
+import static org.mockito.Matchers.argThat;
+import static org.mockito.Mockito.*;
+import static org.mockito.internal.util.collections.Sets.newSet;
+
+/**
+ * @author David BRASSELY (david.brassely at graviteesource.com)
+ * @author GraviteeSource Team
+ */
+public class ApplicationRepositoryMock extends AbstractRepositoryMock<ApplicationRepository> {
+
+    public ApplicationRepositoryMock() {
+        super(ApplicationRepository.class);
+    }
+
+    @Override
+    void prepare(ApplicationRepository applicationRepository) throws Exception {
+        final Application application = mock(Application.class);
+        when(application.getId()).thenReturn("application-sample");
+        when(applicationRepository.findById("application-sample")).thenReturn(of(application));
+
+        final Set<Application> allApplications = newSet(
+                application,
+                mock(Application.class),
+                mock(Application.class),
+                mock(Application.class),
+                mock(Application.class),
+                mock(Application.class),
+                mock(Application.class),
+                mock(Application.class),
+                mock(Application.class));
+        when(applicationRepository.findAll()).thenReturn(allApplications);
+        doAnswer(invocation -> allApplications.remove(application)).when(applicationRepository).delete("deleted-app");
+        when(applicationRepository.findById("deleted-app")).thenReturn(empty());
+
+        final Application newApplication = mock(Application.class);
+        when(newApplication.getName()).thenReturn("created-app");
+        when(newApplication.getDescription()).thenReturn("Application description");
+        when(newApplication.getType()).thenReturn("type");
+        when(newApplication.getStatus()).thenReturn(ApplicationStatus.ACTIVE);
+        when(newApplication.getCreatedAt()).thenReturn(parse("11/02/2016"));
+        when(newApplication.getUpdatedAt()).thenReturn(parse("12/02/2016"));
+
+        when(applicationRepository.findById("created-app")).thenReturn(of(newApplication));
+
+        final Application updatedApplication = mock(Application.class);
+        when(updatedApplication.getId()).thenReturn("updated-app");
+        when(updatedApplication.getName()).thenReturn("updated-app");
+        when(updatedApplication.getDescription()).thenReturn("Updated description");
+        when(updatedApplication.getType()).thenReturn("update-type");
+        when(updatedApplication.getStatus()).thenReturn(ApplicationStatus.ARCHIVED);
+        when(updatedApplication.getCreatedAt()).thenReturn(parse("11/02/2016"));
+        when(updatedApplication.getUpdatedAt()).thenReturn(parse("22/02/2016"));
+
+        when(applicationRepository.findById("updated-app")).thenReturn(of(updatedApplication));
+
+        final Application groupedApplication1 = mock(Application.class);
+        when(groupedApplication1.getId()).thenReturn("grouped-app1");
+        when(groupedApplication1.getGroups()).thenReturn(singleton("application-group"));
+        when(applicationRepository.findById("grouped-app1")).thenReturn(of(groupedApplication1));
+
+        final Application groupedApplication2 = mock(Application.class);
+        when(groupedApplication2.getId()).thenReturn("grouped-app2");
+        when(groupedApplication2.getGroups()).thenReturn(singleton("application-group"));
+        when(applicationRepository.findById("grouped-app2")).thenReturn(of(groupedApplication2));
+
+        final Set<Application> allArchivedApplications = newSet(groupedApplication2);
+        when(applicationRepository.findAll(ApplicationStatus.ARCHIVED)).thenReturn(allArchivedApplications);
+
+        final Application searchedApp1 = mock(Application.class);
+        final Application searchedApp2 = mock(Application.class);
+        when(searchedApp1.getId()).thenReturn("searched-app1");
+        when(searchedApp1.getName()).thenReturn("searched-app1");
+        when(searchedApp2.getId()).thenReturn("searched-app2");
+        when(searchedApp2.getName()).thenReturn("searched-app2");
+        when(applicationRepository.findByName("searched-app1")).thenReturn(singleton(searchedApp1));
+        when(applicationRepository.findByName("arched")).thenReturn(newSet(searchedApp1, searchedApp2));
+        when(applicationRepository.findByName("aRcHEd")).thenReturn(newSet(searchedApp1, searchedApp2));
+
+        when(applicationRepository.findByIds(asList("searched-app1", "searched-app2"))).thenReturn(newSet(searchedApp1, searchedApp2));
+        when(applicationRepository.findByGroups(singletonList("application-group"))).thenReturn(newSet(groupedApplication1, groupedApplication2));
+        when(applicationRepository.findByGroups(singletonList("application-group"), ApplicationStatus.ARCHIVED)).thenReturn(newSet(groupedApplication2));
+
+
+        when(applicationRepository.findByIds(asList("application-sample", "updated-app", "unknown"))).
+                thenReturn(new HashSet<>(asList(application, updatedApplication)));
+
+        when(applicationRepository.update(argThat(new ArgumentMatcher<Application>() {
+            @Override
+            public boolean matches(Object o) {
+                return o == null || (o instanceof Application && ((Application) o).getId().equals("unknown"));
+            }
+        }))).thenThrow(new IllegalStateException());
+
+
+        final Application applicationWithClientId = mock(Application.class);
+        when(applicationWithClientId.getId()).thenReturn("app-with-client-id");
+        when(applicationWithClientId.getClientId()).thenReturn("my-client-id");
+        when(applicationRepository.findByClientId("my-client-id")).thenReturn(of(applicationWithClientId));
+        when(applicationRepository.findByClientId("unknown-client-id")).thenReturn(Optional.empty());
+    }
+}
