+++ src/test/java/io/gravitee/repository/config/mock/EventRepositoryMock.java
@@ -1,0 +1,192 @@
+/**
+ * Copyright (C) 2015 The Gravitee team (http://gravitee.io)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.gravitee.repository.config.mock;
+
+import io.gravitee.repository.management.api.EventRepository;
+import io.gravitee.repository.management.api.search.EventCriteria;
+import io.gravitee.repository.management.api.search.builder.PageableBuilder;
+import io.gravitee.repository.management.model.Event;
+import io.gravitee.repository.management.model.EventType;
+import org.mockito.ArgumentMatcher;
+
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.Map;
+
+import static io.gravitee.repository.utils.DateUtils.parse;
+import static java.util.Arrays.asList;
+import static java.util.Collections.singletonList;
+import static java.util.Optional.empty;
+import static java.util.Optional.of;
+import static org.mockito.Matchers.argThat;
+import static org.mockito.Mockito.*;
+
+/**
+ * @author David BRASSELY (david.brassely at graviteesource.com)
+ * @author GraviteeSource Team
+ */
+public class EventRepositoryMock extends AbstractRepositoryMock<EventRepository> {
+
+    public EventRepositoryMock() {
+        super(EventRepository.class);
+    }
+
+    @Override
+    void prepare(EventRepository eventRepository) throws Exception {
+        final Event event1 = mock(Event.class);
+        final Event event2 = mock(Event.class);
+        final Event event3 = mock(Event.class);
+        final Event event4 = mock(Event.class);
+        final Event event5 = mock(Event.class);
+        final Event event6 = mock(Event.class);
+        final io.gravitee.common.data.domain.Page<Event> pageEvent = mock(io.gravitee.common.data.domain.Page.class);
+        final io.gravitee.common.data.domain.Page<Event> pageEvent2 = mock(io.gravitee.common.data.domain.Page.class);
+        final io.gravitee.common.data.domain.Page<Event> pageEvent3 = mock(io.gravitee.common.data.domain.Page.class);
+        final io.gravitee.common.data.domain.Page<Event> pageEvent4 = mock(io.gravitee.common.data.domain.Page.class);
+        final io.gravitee.common.data.domain.Page<Event> pageEvent5 = mock(io.gravitee.common.data.domain.Page.class);
+        final io.gravitee.common.data.domain.Page<Event> pageEvent6 = mock(io.gravitee.common.data.domain.Page.class);
+        final io.gravitee.common.data.domain.Page<Event> pageEvent7 = mock(io.gravitee.common.data.domain.Page.class);
+        final io.gravitee.common.data.domain.Page<Event> pageEvent8 = mock(io.gravitee.common.data.domain.Page.class);
+        final io.gravitee.common.data.domain.Page<Event> pageEvent9 = mock(io.gravitee.common.data.domain.Page.class);
+
+        Map<String, String> eventProperties = new HashMap<>();
+        eventProperties.put("api_id", "api-1");
+
+        Map<String, String> eventProperties2 = new HashMap<>();
+        eventProperties.put("api_id", "api-4");
+
+        when(event1.getId()).thenReturn("event1");
+        when(event1.getCreatedAt()).thenReturn(parse("11/02/2016"));
+        when(event1.getType()).thenReturn(EventType.PUBLISH_API);
+        when(event1.getPayload()).thenReturn("{}");
+        when(event1.getProperties()).thenReturn(eventProperties);
+        when(event2.getId()).thenReturn("event2");
+        when(event2.getType()).thenReturn(EventType.UNPUBLISH_API);
+        when(event2.getCreatedAt()).thenReturn(parse("12/02/2016"));
+        when(event2.getProperties()).thenReturn(eventProperties);
+        when(event3.getId()).thenReturn("event3");
+        when(event3.getType()).thenReturn(EventType.PUBLISH_API);
+        when(event3.getCreatedAt()).thenReturn(parse("13/02/2016"));
+        when(event4.getId()).thenReturn("event4");
+        when(event4.getType()).thenReturn(EventType.STOP_API);
+        when(event4.getCreatedAt()).thenReturn(parse("14/02/2016"));
+        when(event4.getProperties()).thenReturn(eventProperties2);
+        when(event5.getId()).thenReturn("event5");
+        when(event5.getType()).thenReturn(EventType.START_API);
+        when(event5.getCreatedAt()).thenReturn(parse("15/02/2016"));
+        when(event6.getId()).thenReturn("event6");
+        when(event6.getType()).thenReturn(EventType.START_API);
+        when(event6.getCreatedAt()).thenReturn(parse("16/02/2016"));
+
+        when(eventRepository.findById("event1")).thenReturn(of(event1));
+
+        when(eventRepository.findById("event5")).thenReturn(of(event5), empty());
+
+        when(eventRepository.create(any(Event.class))).thenReturn(event1);
+
+        when(pageEvent.getTotalElements()).thenReturn(2L);
+        when(pageEvent.getContent()).thenReturn(asList(event6, event5));
+        when(eventRepository.search(
+                new EventCriteria.Builder().from(1451606400000L).to(1470157767000L).types(EventType.START_API).build(),
+                new PageableBuilder().pageNumber(0).pageSize(10).build())).thenReturn(pageEvent);
+
+        when(pageEvent2.getTotalElements()).thenReturn(3L);
+        when(pageEvent2.getContent()).thenReturn(asList(event6, event5, event4));
+        when(eventRepository.search(
+                new EventCriteria.Builder().from(1451606400000L).to(1470157767000L).types(EventType.START_API, EventType.STOP_API).build(),
+                new PageableBuilder().pageNumber(0).pageSize(10).build())).thenReturn(pageEvent2);
+
+        when(pageEvent3.getTotalElements()).thenReturn(2L);
+        when(pageEvent3.getContent()).thenReturn(asList(event2, event1));
+        when(eventRepository.search(
+                new EventCriteria.Builder()
+                        .from(1451606400000L).to(1470157767000L)
+                        .property(Event.EventProperties.API_ID.getValue(), "api-1")
+                        .build(),
+                new PageableBuilder().pageNumber(0).pageSize(10).build())).thenReturn(pageEvent3);
+
+        when(pageEvent4.getTotalElements()).thenReturn(0L);
+        when(pageEvent4.getContent()).thenReturn(Collections.emptyList());
+        when(eventRepository.search(
+                new EventCriteria.Builder().from(1420070400000L).to(1422748800000L).types(EventType.START_API).build(),
+                new PageableBuilder().pageNumber(0).pageSize(10).build())).thenReturn(pageEvent4);
+
+        when(pageEvent5.getTotalElements()).thenReturn(1L);
+        when(pageEvent5.getContent()).thenReturn(singletonList(event4));
+        when(eventRepository.search(
+                new EventCriteria.Builder()
+                        .from(1451606400000L).to(1470157767000L)
+                        .property(Event.EventProperties.API_ID.getValue(), "api-3")
+                        .types(EventType.START_API, EventType.STOP_API)
+                        .build(),
+                new PageableBuilder().pageNumber(0).pageSize(10).build())).thenReturn(pageEvent5);
+
+        when(pageEvent6.getTotalElements()).thenReturn(2L);
+        when(pageEvent6.getContent()).thenReturn(asList(event2, event1));
+        when(eventRepository.search(
+                new EventCriteria.Builder()
+                        .from(1451606400000L).to(1470157767000L)
+                        .property(Event.EventProperties.API_ID.getValue(), "api-1")
+                        .build(),
+                null)).thenReturn(pageEvent6);
+
+        when(pageEvent7.getTotalElements()).thenReturn(3L);
+        when(pageEvent7.getContent()).thenReturn(asList(event4, event2, event1));
+        when(eventRepository.search(
+                new EventCriteria.Builder()
+                        .from(1451606400000L).to(1470157767000L)
+                        .property(Event.EventProperties.API_ID.getValue(), asList("api-1", "api-3"))
+                        .build(),
+                null)).thenReturn(pageEvent7);
+
+        when(pageEvent8.getTotalElements()).thenReturn(3L);
+        when(pageEvent8.getPageElements()).thenReturn(2L);
+        when(pageEvent8.getContent()).thenReturn(asList(event6, event2));
+        when(eventRepository.search(
+                new EventCriteria.Builder().from(1451606400000L).to(1470157767000L).types(EventType.START_API, EventType.STOP_API).build(),
+                new PageableBuilder().pageNumber(0).pageSize(2).build())).thenReturn(pageEvent8);
+
+        when(pageEvent9.getTotalElements()).thenReturn(3L);
+        when(pageEvent9.getPageElements()).thenReturn(1L);
+        when(pageEvent9.getContent()).thenReturn(singletonList(event4));
+        when(eventRepository.search(
+                new EventCriteria.Builder().from(1451606400000L).to(1470157767000L).types(EventType.START_API, EventType.STOP_API).build(),
+                new PageableBuilder().pageNumber(1).pageSize(2).build())).thenReturn(pageEvent9);
+
+        when(eventRepository.search(
+                new EventCriteria.Builder()
+                        .from(1451606400000L).to(1470157767000L)
+                        .property(Event.EventProperties.API_ID.getValue(), asList("api-1", "api-3"))
+                        .build())).thenReturn(asList(event4, event2, event1));
+
+        when(eventRepository.search(
+                new EventCriteria.Builder()
+                        .property(Event.EventProperties.API_ID.getValue(), asList("api-1", "api-3"))
+                        .build())).thenReturn(asList(event4, event2, event1));
+
+        when(eventRepository.search(
+                new EventCriteria.Builder().types(EventType.GATEWAY_STARTED).build(),
+                new PageableBuilder().pageNumber(0).pageSize(10).build())).thenReturn(
+                new io.gravitee.common.data.domain.Page<>(Collections.emptyList(), 0, 2, 0));
+
+        when(eventRepository.update(argThat(new ArgumentMatcher<Event>() {
+            @Override
+            public boolean matches(Object o) {
+                return o == null || (o instanceof Event && ((Event) o).getId().equals("unknown"));
+            }
+        }))).thenThrow(new IllegalStateException());
+    }
+}
