+++ src/test/java/io/gravitee/repository/config/mock/MediaRepositoryMock.java
@@ -1,0 +1,83 @@
+/**
+ * Copyright (C) 2015 The Gravitee team (http://gravitee.io)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.gravitee.repository.config.mock;
+
+import io.gravitee.repository.media.api.MediaRepository;
+import io.gravitee.repository.media.model.Media;
+
+import java.util.Date;
+
+import static java.util.Optional.empty;
+import static java.util.Optional.of;
+import static org.mockito.Matchers.any;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
+
+/**
+ * @author David BRASSELY (david.brassely at graviteesource.com)
+ * @author GraviteeSource Team
+ */
+public class MediaRepositoryMock extends AbstractRepositoryMock<MediaRepository> {
+
+    public MediaRepositoryMock() {
+        super(MediaRepository.class);
+    }
+
+    @Override
+    void prepare(MediaRepository mediaRepository) throws Exception {
+        final Media mediaData = mock(Media.class);
+        when(mediaData.getId()).thenReturn("223344");
+        when(mediaData.getFileName()).thenReturn("gravitee_logo_anim.gif");
+        when(mediaData.getSize()).thenReturn(85361L);
+        when(mediaData.getType()).thenReturn("image");
+        when(mediaData.getSubType()).thenReturn("gif");
+        when(mediaData.getHash()).thenReturn("4692FBACBEF919061ECF328CA543E028");
+        when(mediaData.getCreatedAt()).thenReturn(new Date());
+
+
+        final Media mediaData2 = mock(Media.class);
+        when(mediaData2.getId()).thenReturn("556677");
+        when(mediaData2.getFileName()).thenReturn("default_photo.png");
+        when(mediaData2.getSize()).thenReturn(85361L);
+        when(mediaData2.getApi()).thenReturn("123456");
+        when(mediaData2.getHash()).thenReturn("1BC5D9656D860DE678CBEF5C169D8B15");
+
+        final Media mediaData3 = mock(Media.class);
+        when(mediaData3.getId()).thenReturn("22334455");
+        when(mediaData3.getFileName()).thenReturn("stars.png");
+        when(mediaData3.getSize()).thenReturn(4370L);
+        when(mediaData3.getType()).thenReturn("image");
+        when(mediaData3.getSubType()).thenReturn("png");
+        when(mediaData3.getHash()).thenReturn("77C921AB285376AFF72FBDD2D0784E0B");
+        when(mediaData3.getCreatedAt()).thenReturn(new Date());
+
+
+
+        when(mediaRepository.save(any(Media.class))).thenReturn("223344");
+
+        when(mediaRepository.findByHash("4692FBACBEF919061ECF328CA543E028", "image"))
+                .thenReturn(of(mediaData));
+
+        when(mediaRepository.findByHash("77C921AB285376AFF72FBDD2D0784E0B", "123456","image"))
+                .thenReturn(of(mediaData3));
+
+        when(mediaRepository.findByHash("1BC5D9656D860DE678CBEF5C169D8B15", "image"))
+                .thenReturn(of(mediaData2), empty());
+
+        when(mediaRepository.findByHash("1BC5D9656D860DE678CBEF5C169D8B152", "image"))
+                .thenReturn(of(mediaData2), empty());
+    }
+}
