+++ src/test/java/io/gravitee/repository/config/mock/TagRepositoryMock.java
@@ -1,0 +1,71 @@
+/**
+ * Copyright (C) 2015 The Gravitee team (http://gravitee.io)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.gravitee.repository.config.mock;
+
+import io.gravitee.repository.management.api.TagRepository;
+import io.gravitee.repository.management.model.Tag;
+import org.mockito.ArgumentMatcher;
+
+import java.util.Set;
+
+import static java.util.Optional.of;
+import static org.mockito.Matchers.argThat;
+import static org.mockito.Mockito.*;
+import static org.mockito.internal.util.collections.Sets.newSet;
+
+/**
+ * @author David BRASSELY (david.brassely at graviteesource.com)
+ * @author GraviteeSource Team
+ */
+public class TagRepositoryMock extends AbstractRepositoryMock<TagRepository> {
+
+    public TagRepositoryMock() {
+        super(TagRepository.class);
+    }
+
+    @Override
+    void prepare(TagRepository tagRepository) throws Exception {
+        final Tag tag = mock(Tag.class);
+        when(tag.getName()).thenReturn("Tag name");
+        when(tag.getDescription()).thenReturn("Description for the new tag");
+
+        final Tag tag2 = mock(Tag.class);
+        when(tag2.getId()).thenReturn("tag");
+        when(tag2.getName()).thenReturn("Products");
+
+        final Tag tag2Updated = mock(Tag.class);
+        when(tag2Updated.getName()).thenReturn("New product");
+        when(tag2Updated.getDescription()).thenReturn("New description");
+
+        final Set<Tag> tags = newSet(tag, tag2, mock(Tag.class));
+        final Set<Tag> tagsAfterDelete = newSet(tag, tag2);
+        final Set<Tag> tagsAfterAdd = newSet(tag, tag2, mock(Tag.class), mock(Tag.class));
+
+        when(tagRepository.findAll()).thenReturn(tags, tagsAfterAdd, tags, tagsAfterDelete, tags);
+
+        when(tagRepository.create(any(Tag.class))).thenReturn(tag);
+
+        when(tagRepository.findById("new-tag")).thenReturn(of(tag));
+        when(tagRepository.findById("products")).thenReturn(of(tag2), of(tag2Updated));
+
+        when(tagRepository.update(argThat(new ArgumentMatcher<Tag>() {
+            @Override
+            public boolean matches(Object o) {
+                return o == null || (o instanceof Tag && ((Tag) o).getId().equals("unknown"));
+            }
+        }))).thenThrow(new IllegalStateException());
+    }
+}
