+++ src/test/resources/data/user-tests/users.json
@@ -1,56 +1,50 @@
 [
    {
       "id": "user0",
-      "username":"user0 name",
       "email":"user0@gravitee.io",
       "createdAt": 1439022010883,
       "updatedAt": 1439022010883
    },
    {
       "id": "user1",
-      "username":"user1 name",
       "email":"user1@gravitee.io",
+      "source":"gravitee",
+      "sourceId":"user1",
       "createdAt": 1439022010883,
       "updatedAt": 1439022010883
    },
    {
       "id": "user2",
-      "username":"user2 name",
       "email":"user2@gravitee.io",
       "createdAt": 1439022010883,
       "updatedAt": 1439022010883
    },
    {
       "id": "user3",
-      "username":"user3 name",
       "email":"user3@gravitee.io",
       "createdAt": 1439022010883,
       "updatedAt": 1439022010883
    },
    {
       "id": "user4",
-      "username":"user4 name",
       "email":"user4@gravitee.io",
       "createdAt": 1439022010883,
       "updatedAt": 1439022010883
    },
    {
       "id": "user5",
-      "username":"user5 name",
       "email":"user5@gravitee.io",
       "createdAt": 1439022010883,
       "updatedAt": 1439022010883
    },
    {
       "id": "user2delete",
-      "username":"user2delete name",
       "email":"user2delete@gravitee.io",
       "createdAt": 1439022010883,
       "updatedAt": 1439022010883
    },
    {
       "id": "id2update",
-      "username":"username2update",
       "source":"source2update",
       "sourceId":"sourceId2update",
       "password":"password2update",
